Removed OpenFile.h/cpp and merged into ContainerFD. Removed extra references to OpenFile in Container.cpp
